Split publishing and installation into multiple steps
